requirements.txt
pyyaml
requests_oauthlib
